chore(ShoeForm): move form to its own component
